pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ytclfr"
version = "0.1.0"
description = "YouTube video intelligence pipeline"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi",
    "uvicorn",
    "slowapi",
    "celery[redis]",
    "redis",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "faster-whisper",
    "pytesseract",
    "yt-dlp",
    "python-jose[cryptography]",
    "python-multipart",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ytclfr"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
